# Fix `Quad.Contains()` not working with clockwise-oriented quads