Closes https://github.com/ppy/osu/issues/23042

Hopefully the last one of these... I'll let the diff do most of the talking here, as I added a huge explanatory inline comment to attempt to explain what is going on in the new implementation. Scrutiny is however both warranted and welcome.

## Benchmarking

### Before

|   Method | NumPoints |        Mean |     Error |    StdDev | Allocated |
|--------- |---------- |------------:|----------:|----------:|----------:|
| Contains |        10 |    72.03 ns |  1.437 ns |  2.479 ns |         - |
| Contains |       100 |   661.56 ns |  9.191 ns |  8.147 ns |         - |
| Contains |      1000 | 6,723.38 ns | 33.804 ns | 28.228 ns |         - |

### After

|   Method | NumPoints |        Mean |      Error |     StdDev | Allocated |
|--------- |---------- |------------:|-----------:|-----------:|----------:|
| Contains |        10 |    78.49 ns |   1.575 ns |   2.632 ns |         - |
| Contains |       100 |   736.67 ns |  12.492 ns |  16.676 ns |         - |
| Contains |      1000 | 7,688.50 ns | 152.251 ns | 203.250 ns |         - |

Ever so slightly slower, but not by much.